+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
+
+## [Unrelease]
+### Added
+- Delete table, bucket or container
+
 ## [2.1.28] - 2019-09-12
 ### Feature
 - Sync apm configuration with afs portal and os environment (for demo only)